Updated Dockerfile and others
- Added Key files
- Added Sessions
- Added and Changed some Endpoints

+++ server/microservices/stock_arrangement/src/main/kotlin/server/circlehelp/api/response/StockItem.kt
@@ -1,7 +1,9 @@
 package server.circlehelp.api.response
 
+import java.math.BigDecimal
+
 data class StockItem(
     val id: Long,
     val name: String,
-    val price: Double,
+    val price: BigDecimal,
 )
